add playground1,2, docs about project rename something.py to using_mp.py
