+++ .gitignore
@@ -163,4 +163,3 @@ cython_debug/
 # My Own
 /results
 /images
-*/playground.py
