+++ solutions/palm_lines_detector.py
@@ -1,5 +1,5 @@
 import cv2
-import something
+import using_mp
 import utils
 import mediapipe as mp
 import numpy as np
@@ -24,7 +24,7 @@ for i in range(number_of_sample_images):
 
     width = img.shape[1]
     height = img.shape[0]
-    sonnal, coords = something.get_palm(image)
+    sonnal, coords = using_mp.get_palm(img)
 
     # # coords에 있는 값들 중 유효하지 않은 값은 [0,0]이 되도록 설정했습니다.
     # # 만약 그 값을 제외하고 사용하고자 한다면(제 생각에 나름 합리적인 방법인)
@@ -38,14 +38,14 @@ for i in range(number_of_sample_images):
     palm = np.append(sonnal, coords, axis=0)
     # img = cv2.polylines(img, [palm], True, (0, 255, 0), 2)
 
-    landmarks = np.array(something.get_hand_landmark(image))
+    landmarks = np.array(using_mp.get_hand_landmark(img))
 
     boundingRect = cv2.boundingRect(landmarks)
     # cv2.rectangle(img, boundingRect, (0, 0, 255), 3)
-    x1,y1,x2,y2 = boundingRect
+    x1, y1, x2, y2 = boundingRect
     # cv2.circle(img, (x1+ round(x2/2), y1), 5, (255, 255, 0), 10)
 
-    wrist = [landmarks[0][0],landmarks[0][1]]
+    wrist = [landmarks[0][0], landmarks[0][1]]
     # 손바닥 중심
 
     palm_except_fingers = np.array([landmarks[0],   # WRIST
@@ -57,39 +57,39 @@ for i in range(number_of_sample_images):
 
     pivot = utils.center(palm_except_fingers)
 
-    pts = np.array([wrist, pivot],np.int32)
+    pts = np.array([wrist, pivot], np.int32)
 
     # cv2.circle(img, wrist, 5, (0, 0, 255), 2)
     cv2.circle(img, pivot, 5, (0, 0, 255), 2)
 
-    cv2.imshow(f"image{i} original",img)
+    cv2.imshow(f"image{i} original", img)
     # cv2.circle(img, (pivot[0], pivot[1] + 10), 5, (0, 0, 255), 2)
 
     degree_to_rotate = utils.getAngle(pivot, wrist) - 90
 
-    left_top = (0,0)
-    right_top = (width-1, 0)
-    left_bottom = (0,height-1)
-    right_bottom = (width-1, height-1)
+    left_top = (0, 0)
+    right_top = (width - 1, 0)
+    left_bottom = (0, height - 1)
+    right_bottom = (width - 1, height - 1)
 
+    rotated_left_top = utils.rotate_point(left_top, pivot, -degree_to_rotate)
+    rotated_right_top = utils.rotate_point(right_top, pivot, -degree_to_rotate)
+    rotated_left_bottom = utils.rotate_point(
+        left_bottom, pivot, -degree_to_rotate)
+    rotated_right_bottom = utils.rotate_point(
+        right_bottom, pivot, -degree_to_rotate)
 
+    min_x, max_x, _, _ = cv2.minMaxLoc(np.array([rotated_left_top[0],
+                                                rotated_right_top[0],
+                                                rotated_left_bottom[0],
+                                                rotated_right_bottom[0],
+                                                 ]))
 
-    rotated_left_top = utils.rotate_point(left_top,pivot, -degree_to_rotate)
-    rotated_right_top = utils.rotate_point(right_top,pivot, -degree_to_rotate)
-    rotated_left_bottom = utils.rotate_point(left_bottom,pivot, -degree_to_rotate)
-    rotated_right_bottom = utils.rotate_point(right_bottom,pivot, -degree_to_rotate)
-
-    min_x, max_x, _, _= cv2.minMaxLoc(np.array([rotated_left_top[0],
-                                   rotated_right_top[0],
-                                   rotated_left_bottom[0],
-                                   rotated_right_bottom[0],
-                                   ]))
-
-    min_y, max_y, _, _= cv2.minMaxLoc(np.array([rotated_left_top[1],
-                                   rotated_right_top[1],
-                                   rotated_left_bottom[1],
-                                   rotated_right_bottom[1],
-                                   ]))
+    min_y, max_y, _, _ = cv2.minMaxLoc(np.array([rotated_left_top[1],
+                                                rotated_right_top[1],
+                                                rotated_left_bottom[1],
+                                                rotated_right_bottom[1],
+                                                 ]))
     rotated_width, rotated_height = round(max_x - min_x), round(max_y - min_y)
 
     # cv2.putText(img, f"{angle}", (20, 50), cv2.FONT_HERSHEY_SIMPLEX, 1, (255, 0, 255))
@@ -97,10 +97,9 @@ for i in range(number_of_sample_images):
 
     M = cv2.getRotationMatrix2D(pivot, degree_to_rotate, 1.0)
 
-
-    top_border_size= min_y * -1 if min_y < 0 else 0
+    top_border_size = min_y * -1 if min_y < 0 else 0
     bottom_border_size = max_y - width if max_y - width > 0 else 0
-    left_border_size= min_x * -1 if min_x < 0 else 0
+    left_border_size = min_x * -1 if min_x < 0 else 0
     right_border_size = max_x - width if max_x - width > 0 else 0
 
     horizontal_border_size = round((rotated_width - width) / 2)
@@ -121,8 +120,6 @@ for i in range(number_of_sample_images):
     # rotated_img = cv2.CreateMat(rotated_height, rotated_width, )
     img = cv2.warpAffine(img, M, (rotated_width, rotated_height))
 
-
-
     # 아래 이 주석은 수직, 수평 테두리 두께를 출력하는 코드임
     hori_text = f"hori: {horizontal_border_size}"
     vert_text = f"vert: {vertical_border_size}"
@@ -133,8 +130,6 @@ for i in range(number_of_sample_images):
     # cv2.putText(img, vert_text, (50,50+text_size[0][1]+10),cv2.FONT_HERSHEY_SIMPLEX ,
     #                1, (0,255,0), 2, cv2.LINE_AA)
 
-
-
     # print(text_size[0])
     # cv2.putText(img, vert_text, (50+text_size[0],50+text_size[1]),cv2.FONT_HERSHEY_SIMPLEX ,
     #                1, (0,255,0), 2, cv2.LINE_AA)
@@ -144,14 +139,14 @@ for i in range(number_of_sample_images):
     # for coord in landmarks:
     #     cv2.circle(img, coord, 5, (0, 0, 255), 2)
 
-    cv2.imshow(f"image{i}",img)
+    cv2.imshow(f"image{i}", img)
 
     k = cv2.waitKey(0)
     cv2.destroyAllWindows()
     # for문 도중 Esc를 누르면 프로그램이 종료되게 함
-    if k==27:    # Esc key to stop
-         break
-    elif k==-1:  # normally -1 returned,so don't print it
-         continue
+    if k == 27:    # Esc key to stop
+        break
+    elif k == -1:  # normally -1 returned,so don't print it
+        continue
 
 cv2.destroyAllWindows()
